Refactor: Move user-related data classes to data/user package
- Moved User, UserRepository, Gender, Role, BankCode, and RoleType to the `com.example.baytro.data.user` package.
- Updated imports across multiple files to reflect the new package structure.
- Added new fields `contractNumber`, `landlordId`, `buildingId` to `Contract` data class.
- Added `PENDING` and `ENDED` to `Status` enum in `Contract.kt`.
- Updated `SubmitButton` to accept a modifier.
- Adjusted padding in `NewTenantUserScreen` and `NewLandlordUserScreen` for `DividerWithSubhead`.
- Modified `DividerWithSubhead` to accept modifier as the first parameter.
