+++ app/src/main/java/com/example/baytro/data/contract/Contract.kt
@@ -7,7 +7,8 @@ import kotlinx.serialization.Serializable
 enum class Status {
     ACTIVE,
     OVERDUE,
-    TERMINATED
+    PENDING,
+    ENDED
 }
 
 @Serializable
@@ -15,8 +16,11 @@ data class Contract(
     @kotlinx.serialization.Transient
     @DocumentId
     val id: String = "",
+    val contractNumber: String,
+    val landlordId: String,
     val tenantId: List<String>,
     val roomId: String,
+    val buildingId: String,
     val startDate: String,
     val endDate: String,
     val rentalFee: Int,
