+++ app/src/main/java/com/example/baytro/data/user/UserRepository.kt
@@ -1,5 +1,6 @@
-package com.example.baytro.data
+package com.example.baytro.data.user
 
+import com.example.baytro.data.Repository
 import dev.gitlive.firebase.firestore.FirebaseFirestore
