+++ app/src/main/java/com/example/baytro/service/FptAiService.kt
@@ -3,7 +3,7 @@ package com.example.baytro.service
 import android.util.Log
 import com.example.baytro.BuildConfig
 import com.example.baytro.data.FptAiOcrResponse
-import com.example.baytro.data.Gender
+import com.example.baytro.data.user.Gender
 import com.example.baytro.data.IdCardInfo
 import com.example.baytro.data.MediaRepository
 import io.ktor.client.HttpClient
