+++ app/src/main/java/com/example/baytro/view/components/SubmitButton.kt
@@ -1,7 +1,5 @@
 package com.example.baytro.view.components
 
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
@@ -13,16 +11,15 @@ import androidx.compose.ui.unit.dp
 
 @Composable
 fun SubmitButton(
+    modifier: Modifier = Modifier,
     text: String = "Submit",
     isLoading: Boolean,
-    onClick: () -> Unit
+    onClick: () -> Unit,
 ) {
     Button(
         onClick = { if (!isLoading) onClick() },
         enabled = !isLoading,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(50.dp)
+        modifier = modifier
     ) {
         if (isLoading) {
             CircularProgressIndicator(
