+++ app/src/main/java/com/example/baytro/viewModel/auth/SignInVM.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.baytro.auth.AuthRepository
 import com.example.baytro.auth.SignInFormState
-import com.example.baytro.data.UserRepository
+import com.example.baytro.data.user.UserRepository
 import com.example.baytro.utils.ValidationResult
 import com.example.baytro.utils.Validator
 import com.example.baytro.view.AuthUIState
