+++ app/src/main/java/com/example/baytro/viewModel/splash/UploadIdCardVM.kt
@@ -6,7 +6,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.baytro.auth.AuthRepository
-import com.example.baytro.data.Gender
+import com.example.baytro.data.user.Gender
 import com.example.baytro.data.IdCardInfo
 import com.example.baytro.data.IdCardInfoWithImages
 import com.example.baytro.data.MediaRepository
